docs(subagents): address second-pass PR #1581 review feedback
- Document the abnormal-tool-end caveat at both isToolActive callsites:
  if a tool never emits its end event the idle watchdog is deferred
  indefinitely and the wall-clock cap is the sole backstop; add a
  foreground test proving the wall cap terminates such attempts.
- Document that ATOMIC_SUBAGENT_ATTEMPT_KILL_GRACE_MS intentionally
  cannot be disabled (0/negative/non-numeric fall back to the default)
  in the resolver, subagents.md, and both changelogs.
- Note in the classifier conformance suite that it enforces behavioral
  (not structural) parity over a finite corpus, that shared-module
  extraction is blocked by the package split, and that both copies plus
  a fixture must be updated together; add a fixture documenting the
  known benign-"too large" message false positive.
- Split the watchdog test suite (shared helpers + separate pre-spawn
  candidate-filtering suite) to stay under the 500-line file gate.

Refs: #1581
Assistant-model: Claude Fable 5
