+++ test/unit/model-fallback-classifier-conformance.test.ts
@@ -15,6 +15,14 @@ import {
 // suite runs a shared corpus of failure fixtures through both and asserts they
 // agree, so a change to one copy that silently diverges the other fails here.
 //
+// Limitations: this enforces BEHAVIORAL parity over a finite corpus only — the two
+// copies are structurally different (e.g. the subagents copy combines regex
+// alternations that the workflows copy keeps separate), and a divergence on an
+// input outside this corpus would pass CI silently. When adding or changing a
+// classification rule, update BOTH copies and add a fixture here that exercises
+// the new rule. (Extracting a single shared module is blocked by the mandated
+// package split; the copies must not import each other.)
+//
 // Known intentional difference (not covered by the shared corpus): the workflows
 // classifier has an extra `transport_error` kind for bare "connection error." /
 // "fetch failed." wrapper messages that the subagents classifier does not model.
@@ -54,6 +62,14 @@ const CONFORMANCE_FIXTURES: readonly Fixture[] = [
   { label: "message parameter not supported", failure: new Error("parameter not supported by this model"), kind: "request_incompatible", retryable: true },
   { label: "message invalid request", failure: new Error("invalid request"), kind: "request_incompatible", retryable: true },
   { label: "message bad request", failure: { message: "400 bad request" }, kind: "request_incompatible", retryable: true },
+  // Known residual fragility of message-substring classification (documented on
+  // purpose): a benign task error that merely contains "too large" is classified
+  // as retryable request_incompatible by BOTH copies when no structured
+  // status/code and no task-failure/refusal keywords are present. Structured
+  // signals and refusal/cancel/task-failure patterns are checked first, which
+  // bounds the blast radius; if either copy tightens the bare "too large"
+  // pattern, update this fixture in lockstep.
+  { label: "benign message containing too large", failure: new Error("generated diff is too large to display inline"), kind: "request_incompatible", retryable: true },
   // Refusals/cancellations must win over request-incompatible wrappers.
   { label: "aborted stopReason wrapper", failure: { status: 400, stopReason: "aborted", errorMessage: "aborted" }, kind: "cancelled", retryable: false },
   { label: "AbortError name wrapper", failure: { statusCode: 422, name: "AbortError", message: "aborted by user" }, kind: "cancelled", retryable: false },
